+++ python-omegaclient/omegaclient/cluster.py
@@ -13,70 +13,100 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
-from omegaclient.utils import url_maker 
+import webob
+from jsonschema import SchemaError, ValidationError, validate
+from omegaclient.utils import url_maker
+
 
 class ClusterAPI(object):
-      
-    path = "/clusters"
 
-    def __init__(self, http):
-        self.http = http
-
-    def index(self):
+    def get_clusters(self):
         """List all clusters"""
-        return self.http.get(self.path)
 
-    def show(self, id):
-        """List specified cluster"""
-        return self.http.get(url_maker(self.path, id))
+        return self.http.get("/clusters")
 
-    def create(self, body):
-        """Create new cluster"""
-        return self.http.post(self.path, body=body)
+    def get_cluster(self, cluster_id):
+        """List single cluster's information"""
+
+        return self.http.get(url_maker("/clusters", cluster_id))
+
+    def create_cluster(self, name, cluster_type, group_id):
+        """Create new cluster.
+
+        :param name: cluster name
+        :param cluster_type: cluster type, current support values are:
+                               1_master, 3_masters, 5_masters
+        :param group_id: the group which the new cluster belong to
+        """
+        data = {
+              "name": name,
+              "clusterType": cluster_type,
+              "groupId": group_id,
+        }
+        schema = {
+              "type": "object",
+              "properties": {
+                  "name": {"type": "string"},
+                  "clusterType": {"type": "string"},
+                  "groupId": {"type": "number"},
+              },
+              "required": ["name", "clusterType", "groupId"]
+          }
+        try:
+            validate(data, schema)
+        except (SchemaError, ValidationError):
+            raise webob.exc.HTTPBadRequest(explanation="Bad Paramaters")
+
+        return self.http.post("/clusters", data=data)
+
+    def delete_cluster(self, cluster_id):
+        """Delete cluster"""
+
+        return self.http.delete(url_maker("/clusters", cluster_id))
+
+    def update_cluster(self, cluster_id, **kwargs):
+        """Updated cluster information."""
+
+        return self.http.patch(url_maker("/clusters", cluster_id), data=kwargs)
+
+    def get_node_identifier(self, cluster_id):
+        """Generated new node identifier. this identifier will be
+        used for add new node."""
+
+        return self.http.get(url_maker("/clusters", cluster_id,
+                                       "new_node_identifier"))
+
+    def get_cluster_node(self, cluster_id, node_id):
+        """List node information"""
+
+        return self.http.get(url_maker("/clusters", cluster_id, "nodes",
+                                       node_id))
+
+    def update_cluster_node(self, cluster_id, node_id, **kwargs):
+        """Updated node information"""
+
+        return self.http.patch(url_maker("/clusters", cluster_id,
+                                         "nodes", node_id), **kwargs)
+
+    def get_node_metrics(self, cluster_id, node_id):
+        """Retrive node metrics"""
+
+        return self.http.get(url_maker("/clusters", cluster_id, "nodes",
+                                       node_id))
+
+    def update_node_service(self, cluster_id, node_id, service_name, **kwargs):
+        """Reset or restart service on node"""
+
+        return self.http.patch(url_maker("/clusters", cluster_id, "nodes",
+                                         node_id, "services", service_name),
+                               **kwargs)
 
     def create_node(self, cluster_id, **kwargs):
         """Add new node for cluster identified by `cluster_id`"""
-        return self.http.post(url_maker(self.path, cluster_id, "nodes"),
-                              body=kwargs)
-
-    def delete(self, cluster_id):
-        """Delete specified cluster"""
-        return self.http.delete(url_maker(self.path, cluster_id))
+        return self.http.post(url_maker("/clusters", cluster_id, "nodes"),
+                              data=kwargs)
 
     def delete_nodes(self, cluster_id, *args):
         """Delete nodes for cluster identified by `cluster_id`"""
-        return self.http.delete(url_maker(self.path, cluster_id, "node"),
-                                body=args)
-
-    def update(self,  body):
-        """Update cluster partial"""
-        return self.http.patch(url_maker(self.path, id), body=body)
-    
-    def versions(self):
-        pass
-
-    def identifier(self, cluster_id):
-        """Generated new node uuid belong to specified cluster"""
-        return self.http.get(url_maker(self.path, cluster_id,
-                                       "new_node_identifier"))
-    
-    def node(self, cluster_id, node_id):
-        """List node information for specified cluster"""
-        return self.http.get(url_maker(self.path, cluster_id, "nodes",
-                                       node_id))
-    
-    def update_node(self, cluster_id, node_id, **kwargs):
-        """Updated node information for speicified cluster"""
-        return self.http.patch(url_maker(self.path, cluster_id,
-                                         "nodes", node_id), **kwargs)
-
-    def metrics(self, cluster_id, node_id):
-        """List metrics for node of specified cluster"""
-        return self.http.get(url_maker(self.path, cluster_id, "nodes",
-                                       node_id))
-
-    def service(self, cluster_id, node_id, service_name, **kwargs):
-        """Reset or Restart specified service"""
-        return self.http.patch(url_maker(self.path, cluster_id, "nodes",
-                                         node_id, "services", service_name),
-                               **kwargs)
+        return self.http.delete(url_maker("/clusters", cluster_id, "node"),
+                                data=args)
